Add files via upload
Background:

Home Credit uses various statistical methods and Machine Learning to make credit score predictions. Companies ask to unlock the maximum potential of their data. Doing so will ensure that clients capable of repayment are accepted and that loans are given with a principal, maturity, and repayment calendar that will empower their clients to be successful.

Objective:

1) Identify customers with characteristics of potential clients who have difficulty repaying loans.
2) Predict the client’s repayment abilities.

Actions:

1) Perform data cleaning, and visualization for business insights.
2) Build a models with machine learning algorithms.
3) Predict customer's repayment abilities on application data test .
4) Provide recommendations for company to increase their clients succeed in applying for loans.